* origin/wlan-cmn.driver.lnx.1.0-dev:
  Release 5.1.0.22D
  qcacmn: Address 32 bit specific compilation issues
  Release 5.1.0.22C
  qcacmn: CL 1546483 - update fw common interface files
  Release 5.1.0.22B
  Release 5.1.0.22A
  qcacmn: Fix incorrect channel list passed to the firmware
  qcacmn: Add workqueue APIs in QDF

+++ wmi/src/wmi_unified_api.c
@@ -3360,7 +3360,7 @@ QDF_STATUS wmi_unified_roam_scan_offload_scan_period(void *wmi_hdl,
  */
 QDF_STATUS wmi_unified_roam_scan_offload_chan_list_cmd(void *wmi_hdl,
 				   uint8_t chan_count,
-				   uint8_t *chan_list,
+				   uint32_t *chan_list,
 				   uint8_t list_type, uint32_t vdev_id)
 {
 	wmi_unified_t wmi_handle = (wmi_unified_t) wmi_hdl;
